fix: chromatic build issues with import mixin
