+++ package.json
@@ -110,16 +110,16 @@
     "eslint-plugin-testing-library": "^5.9.1",
     "mkdirp": "^1.0.4",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.1.7",
+    "postcss": "^8.4.18",
     "postcss-scss": "^4.0.5",
-    "postcss-loader": "^4.0.4",
+    "postcss-loader": "^7.0.1",
     "prettier": "^2.7.1",
     "react-scripts": "^5.0.1",
-    "rollup": "^2.56.3",
+    "rollup": "^3.2.5",
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-typescript2": "^0.31.2",
+    "rollup-plugin-typescript2": "^0.34.1",
     "sass": "^1.56.0",
     "sass-loader": "10.3.1",
     "semantic-release": "^19.0.5",
