+++ src/stories/components/GridPopoutEditGeneric.stories.tsx
@@ -1,6 +1,7 @@
 import "@linzjs/lui/dist/scss/base.scss";
 import "@linzjs/lui/dist/fonts";
 import "../../styles/index.scss";
+import "../../styles/GridTheme.scss";
 
 import { ComponentMeta, ComponentStory } from "@storybook/react/dist/ts3.9/client/preview/types-6-3";
 import { GridUpdatingContextProvider } from "@contexts/GridUpdatingContextProvider";
