+++ src/styles/GridTheme.scss
@@ -1,6 +1,10 @@
-@use "../../node_modules/@linzjs/lui/dist/scss/Foundation/Variables/ColorVars" as lui;
-@import url("../../node_modules/ag-grid-community/src/styles/ag-grid");
-@import url("../../node_modules/ag-grid-community/src/styles/ag-theme-alpine/sass/ag-theme-alpine-mixin");
+@use "@linzjs/lui/dist/scss/Foundation/Variables/ColorVars" as lui;
+
+// These break if we use urls so ignore stylelint
+/* stylelint-disable-next-line */
+@import "ag-grid-community/src/styles/ag-grid";
+/* stylelint-disable-next-line */
+@import "ag-grid-community/src/styles/ag-theme-alpine/sass/ag-theme-alpine-mixin";
 
 .ag-theme-alpine {
   // AG Grid themes can be customized using theme parameters, which are arguments that change its appearance. Refer to
@@ -11,6 +15,7 @@
       alpine-active-color: lui.$sea,
       foreground-color: lui.$grey-80,
       data-color: lui.$grey-80,
+      // ag-grid-community >=28
       selected-row-background-color: lui.$polar,
       header-foreground-color: lui.$surfie,
       background-color: lui.$white,
@@ -43,6 +48,12 @@
 
   .ag-row:last-of-type {
     border-bottom: none;
+  }
+
+  // ag-grid-community 27.x.x needs this vs 28 using the selected-row-background-color above
+  // You can set a background _color_ with theme parameters, but for something like a gradient, you need to override CSS.
+  .ag-row-selected {
+    background-color: lui.$polar;
   }
 
   .ag-header-cell {
